Fix headless mode by outsourcing GUI
